patch: mruby-float-patch: Just make it work for current MRuby
Numbers of users feedback that they all failed to patch
`mruby-float-patch.patch`, failing at mruby/src/backtrace.c.

This is because MRuby updated their APIs, and nowhere to apply patch. To
deal with it, I analyzed this patch: it meant to add a buffer called
`superhack`, and should be a workaround for Zest's mruby-widget-ui gem.
Then I tried to edit this patch to fit current MRuby's version so that
the build process won't fail again.

Currently it doesn't appear to cause other bugs on my successful build.