Remove submodules add docker builds
Adjust new build scripts to be similar to the old ruby ones

- Changes binary names
- Default to demo build
- Fully scripted builder with ubuntu
- Tested build with arch/alpine
- Moved the run zyn in a container script (right now I think it's
  broken)
- Added some quick readmes to explain subdirectories
- Hacked in a partial dirty build mode (ideal for the cases where you
  _almost_ have something building)
- Changes version to git head and version ID to 3.0.6-git

With some other dockerized scripts out there on the net I'm optimistic
that it might be possible to add in a MacOS cross-build, but we'll have
to see how tricky it is. If nothing else it fits into the new workflow
quite well.